Separate factories from command execution
